po: Fixes to the German translation
--

Justus reported that the German translation for the key listing in
--edit-key does "usage: E" -> "Aufruf: E" which is clearly wrong.  It
turnd out that this translation was once marked as fuzzy and
accidentally unfuzzied by me.

   "Aufruf" (bug) -> "Nutzung"
   "Leistungsfähigkeit" -> "Nutzung"
   "Signaturfähigkeit" -> "Signaturnutzbarkeit" etc.

The last two are in the key generation menu.  Also changed the key
code for "Umschalten der Signaturnutzbarkeit" from "U" to "S".

"Nutzung" is here better than "Fähigkeit" because the latter is more
connected to the property of the algorithm, where the former better
expresses an arbitrary choice.

Signed-off-by: Werner Koch <wk@gnupg.org>